refactor(cliproxy): extract plan validation and HTTP probes into setup/validation.ts
Move MODEL_ID_RE, validateSetupOptions, verifyModelsAvailable,
assertProxyReachable, and assertProxyKeyWorks out of setup.ts. The two
assertProxy* helpers become exported so the orchestrator (Unit 9) can
dependency-inject them.

setup.ts shrinks from 1261 to 1114 LOC. Adds 4 new TDD-first tests for
assertProxyReachable/assertProxyKeyWorks happy + error paths (714 total, +16).
