+++ packages/cli/src/commands/cliproxy/setup.test.ts
@@ -137,26 +137,6 @@ const MISSING_OPENCODE_CONFIG_OPENAI_MODEL_WORKFLOW = `      - uses: fro-bot/age
 `
 
 describe('cliproxy setup helpers', () => {
-  describe('validateSetupOptions', () => {
-    it('requires --key in non-interactive mode', () => {
-      expect(() => validateSetupOptions({repo: 'owner/repo', harness: 'opencode'}, false)).toThrow(
-        '--key is required when stdin is not a TTY',
-      )
-    })
-
-    it('requires --repo in non-interactive mode', () => {
-      expect(() => validateSetupOptions({key: 'sk-test', harness: 'opencode'}, false)).toThrow(
-        '--repo is required when stdin is not a TTY',
-      )
-    })
-
-    it('requires --harness in non-interactive mode', () => {
-      expect(() => validateSetupOptions({key: 'sk-test', repo: 'owner/repo'}, false)).toThrow(
-        '--harness is required when stdin is not a TTY',
-      )
-    })
-  })
-
   describe('analyzeFroBotWorkflow', () => {
     it('returns empty stepsWithGaps when all four inputs are wired', () => {
       const result = analyzeFroBotWorkflow(COMPLETE_WORKFLOW)
@@ -460,164 +440,6 @@ describe('option parsing', () => {
   })
 })
 
-describe('verifyModelsAvailable', () => {
-  // Realistic fixture matching the plan spec
-  const MODELS_FIXTURE = {
-    data: [
-      {id: 'claude-3-7-sonnet-20250219', owned_by: 'anthropic'},
-      {id: 'claude-sonnet-4-6', owned_by: 'anthropic'},
-      {id: 'gpt-5.4-mini', owned_by: 'openai'},
-      {id: 'gpt-5.5', owned_by: 'openai'},
-    ],
-  }
-
-  const BASE_URL = 'https://cliproxy.fro.bot'
-  const KEY = 'sk-test-key'
-
-  // Save and restore globalThis.fetch around each test
-  let originalFetch: typeof globalThis.fetch
-  afterEach(() => {
-    globalThis.fetch = originalFetch
-  })
-  // Capture original before any test runs
-  originalFetch = globalThis.fetch
-
-  it('anthropic-only short-circuit: returns immediately without calling fetch', async () => {
-    const fetchSpy = mock(async () => new Response(JSON.stringify(MODELS_FIXTURE)))
-    globalThis.fetch = fetchSpy as unknown as typeof fetch
-
-    await verifyModelsAvailable(BASE_URL, KEY, ['anthropic'], 'anthropic/claude-sonnet-4-6')
-
-    expect(fetchSpy.mock.calls.length).toBe(0)
-  })
-
-  it('happy path: openai-only, model present, owned_by openai — passes without throw', async () => {
-    globalThis.fetch = mock(async () => new Response(JSON.stringify(MODELS_FIXTURE))) as unknown as typeof fetch
-
-    await expect(verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')).resolves.toBeUndefined()
-  })
-
-  it('happy path: dual providers, anthropic model present, openai entries exist — passes', async () => {
-    globalThis.fetch = mock(async () => new Response(JSON.stringify(MODELS_FIXTURE))) as unknown as typeof fetch
-
-    await expect(
-      verifyModelsAvailable(BASE_URL, KEY, ['anthropic', 'openai'], 'anthropic/claude-sonnet-4-6'),
-    ).resolves.toBeUndefined()
-  })
-
-  it('error path: 401 throws "Proxy key rejected" message', async () => {
-    globalThis.fetch = mock(async () => new Response('Unauthorized', {status: 401})) as unknown as typeof fetch
-
-    await expect(verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')).rejects.toThrow(
-      'Proxy key rejected',
-    )
-  })
-
-  it('error path: 401 error message does NOT contain the Authorization header value', async () => {
-    globalThis.fetch = mock(async () => new Response('Unauthorized', {status: 401})) as unknown as typeof fetch
-
-    let errorMessage = ''
-    try {
-      await verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')
-    } catch (error) {
-      errorMessage = error instanceof Error ? error.message : String(error)
-    }
-
-    expect(errorMessage).not.toContain(KEY)
-    expect(errorMessage).not.toContain('Bearer')
-  })
-
-  it('error path: 403 throws "Proxy key rejected" message', async () => {
-    globalThis.fetch = mock(async () => new Response('Forbidden', {status: 403})) as unknown as typeof fetch
-
-    await expect(verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')).rejects.toThrow(
-      'Proxy key rejected',
-    )
-  })
-
-  it('error path: 500 throws with status and truncated body; no Authorization header in message', async () => {
-    const body = 'Internal Server Error — something went wrong on the proxy'
-    globalThis.fetch = mock(async () => new Response(body, {status: 500})) as unknown as typeof fetch
-
-    let errorMessage = ''
-    try {
-      await verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')
-    } catch (error) {
-      errorMessage = error instanceof Error ? error.message : String(error)
-    }
-
-    expect(errorMessage).toContain('500')
-    expect(errorMessage).not.toContain(KEY)
-    expect(errorMessage).not.toContain('Bearer')
-  })
-
-  it('error path: 200 with data:[] and openai in providers throws no-openai-models message', async () => {
-    globalThis.fetch = mock(async () => new Response(JSON.stringify({data: []}))) as unknown as typeof fetch
-
-    await expect(verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')).rejects.toThrow(
-      'No OpenAI models on proxy',
-    )
-  })
-
-  it('error path: model not present in data — throws and lists available openai ids', async () => {
-    globalThis.fetch = mock(async () => new Response(JSON.stringify(MODELS_FIXTURE))) as unknown as typeof fetch
-
-    let errorMessage = ''
-    try {
-      await verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-99-unknown')
-    } catch (error) {
-      errorMessage = error instanceof Error ? error.message : String(error)
-    }
-
-    expect(errorMessage).toContain('gpt-99-unknown')
-    // Should list available openai models
-    expect(errorMessage).toContain('gpt-5.4-mini')
-    expect(errorMessage).toContain('gpt-5.5')
-    // Should NOT list anthropic models
-    expect(errorMessage).not.toContain('claude')
-  })
-
-  it('error path: model not present and provider is anthropic — lists available anthropic ids', async () => {
-    globalThis.fetch = mock(async () => new Response(JSON.stringify(MODELS_FIXTURE))) as unknown as typeof fetch
-
-    let errorMessage = ''
-    try {
-      await verifyModelsAvailable(BASE_URL, KEY, ['anthropic', 'openai'], 'anthropic/claude-unknown-model')
-    } catch (error) {
-      errorMessage = error instanceof Error ? error.message : String(error)
-    }
-
-    expect(errorMessage).toContain('claude-unknown-model')
-    // Should list available anthropic models
-    expect(errorMessage).toContain('claude-3-7-sonnet-20250219')
-    expect(errorMessage).toContain('claude-sonnet-4-6')
-    // Should NOT list openai models
-    expect(errorMessage).not.toContain('gpt-')
-  })
-
-  it('error path: data is missing (response is {}) — throws clean error', async () => {
-    globalThis.fetch = mock(async () => new Response(JSON.stringify({}))) as unknown as typeof fetch
-
-    await expect(verifyModelsAvailable(BASE_URL, KEY, ['openai'], 'openai/gpt-5.4-mini')).rejects.toThrow(
-      /data.*array|unexpected.*response/i,
-    )
-  })
-
-  it('error path: dual providers, no owned_by=openai entries — throws no-openai-models message', async () => {
-    const anthropicOnlyData = {
-      data: [
-        {id: 'claude-3-7-sonnet-20250219', owned_by: 'anthropic'},
-        {id: 'claude-sonnet-4-6', owned_by: 'anthropic'},
-      ],
-    }
-    globalThis.fetch = mock(async () => new Response(JSON.stringify(anthropicOnlyData))) as unknown as typeof fetch
-
-    await expect(verifyModelsAvailable(BASE_URL, KEY, ['anthropic', 'openai'], 'openai/gpt-5.4-mini')).rejects.toThrow(
-      'No OpenAI models on proxy',
-    )
-  })
-})
-
 describe('validation matrix + non-interactive plan', () => {
   const MODELS_FIXTURE = {
     data: [
@@ -636,77 +458,6 @@ describe('validation matrix + non-interactive plan', () => {
     globalThis.fetch = originalFetch
   })
   originalFetch = globalThis.fetch
-
-  // ── validateSetupOptions ──────────────────────────────────────────────────
-
-  describe('validateSetupOptions — providers/model validation', () => {
-    it('regression: no providers/model passes unchanged (anthropic-only default)', () => {
-      expect(() => validateSetupOptions({key: 'sk-test', repo: 'owner/repo', harness: 'opencode'}, false)).not.toThrow()
-    })
-
-    it('happy path: single provider anthropic, no model — passes', () => {
-      expect(() =>
-        validateSetupOptions({key: 'sk-test', repo: 'owner/repo', harness: 'opencode', providers: 'anthropic'}, false),
-      ).not.toThrow()
-    })
-
-    it('happy path: openai + model with openai prefix — passes', () => {
-      expect(() =>
-        validateSetupOptions(
-          {key: 'sk-test', repo: 'owner/repo', harness: 'opencode', providers: 'openai', model: 'openai/gpt-5.4-mini'},
-          false,
-        ),
-      ).not.toThrow()
-    })
-
-    it('happy path: anthropic,openai + model with openai prefix — passes', () => {
-      expect(() =>
-        validateSetupOptions(
-          {
-            key: 'sk-test',
-            repo: 'owner/repo',
-            harness: 'opencode',
-            providers: 'anthropic,openai',
-            model: 'openai/gpt-5.4-mini',
-          },
-          false,
-        ),
-      ).not.toThrow()
-    })
-
-    it('error: multiple providers without --model throws "Pass --model" error', () => {
-      expect(() => validateSetupOptions({harness: 'opencode', providers: 'anthropic,openai'}, false)).toThrow(
-        'Pass --model <provider/model-id> when selecting multiple providers.',
-      )
-    })
-
-    it('error: model prefix does not match single provider (anthropic provider, openai model)', () => {
-      expect(() =>
-        validateSetupOptions({harness: 'opencode', providers: 'anthropic', model: 'openai/gpt-5.4-mini'}, false),
-      ).toThrow(/Model prefix openai does not match selected providers/)
-    })
-
-    it('error: model prefix does not match single provider (openai provider, anthropic model)', () => {
-      expect(() =>
-        validateSetupOptions({harness: 'opencode', providers: 'openai', model: 'anthropic/claude-sonnet-4-6'}, false),
-      ).toThrow(/Model prefix anthropic does not match selected providers/)
-    })
-
-    it('error: duplicate providers throws from parseProviders', () => {
-      expect(() => validateSetupOptions({harness: 'opencode', providers: 'anthropic,anthropic'}, false)).toThrow(
-        /duplicate/,
-      )
-    })
-
-    it('error: unknown provider throws from parseProviders', () => {
-      expect(() => validateSetupOptions({harness: 'opencode', providers: 'claude'}, false)).toThrow(/Unknown provider/)
-    })
-
-    it('interactive mode: providers/model checks are skipped even with invalid combo', () => {
-      // Multiple providers without model — would fail in non-interactive, but interactive skips all checks
-      expect(() => validateSetupOptions({providers: 'anthropic,openai'}, true)).not.toThrow()
-    })
-  })
 
   // ── buildNonInteractivePlan ───────────────────────────────────────────────
